Do not configure bagpipe bgpvpn extension for sriov agent
Currently l2 agent extensions are specified in a single pillar
and then used by ovs and sriov agents.
This is not correct since there might be different extensions for
different agents, so we need separate pillars for each agent.

This patch is a hot fix for the release.
Another bug will be filed for proper extensions handling.

Change-Id: I84111b419a4386e4598edac100cfd746ec7a3611
Related-Prod: PROD-19316